# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizzHive Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login (separate from buyer/seller auth)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginBody"
      responses:
        "200":
          description: Logged in
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Invalid credentials
  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Current admin
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
...
  /admin/overview:
    get:
      operationId: getAdminOverview
      tags: [admin]
      summary: Queue counters for the admin console
      responses:
        "200":
          description: Counters
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOverview"
        "401":
          description: Not authenticated
  /admin/disputes:
    get:
      operationId: listDisputes
      tags: [admin]
      summary: List disputed orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, resolved, all]
      responses:
        "200":
          description: Disputed orders
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DisputeOrder"
        "401":
          description: Not authenticated
  /admin/disputes/{id}:
    get:
      operationId: getDispute
      tags: [admin]
      summary: Get a single disputed order
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Disputed order
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DisputeOrder"
        "404":
          description: Order not found
  /admin/disputes/{id}/resolve:
    post:
      operationId: resolveDispute
      tags: [admin]
      summary: Resolve a dispute by releasing funds or refunding the buyer
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResolveDisputeBody"
      responses:
        "200":
          description: Dispute resolved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResolveDisputeResponse"
        "400":
          description: Invalid resolution or order not under dispute
        "502":
          description: Refund could not be processed
  /admin/support-tickets:
    get:
      operationId: listSupportTickets
      tags: [admin]
      summary: Support queue, Premium vendors first
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, closed, all]
      responses:
        "200":
          description: Support tickets
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/SupportTicket"
  /admin/support-tickets/{id}/close:
    post:
      operationId: closeSupportTicket
      tags: [admin]
      summary: Close a support ticket
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CloseSupportTicketBody"
      responses:
        "200":
          description: Ticket closed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CloseSupportTicketResponse"
...
components:
  schemas:
...
    SupportTicket:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        subject:
          type: string
        message:
          type: string
        requesterRole:
          type: string
        userId:
          type: ["integer", "null"]
        vendorId:
          type: ["integer", "null"]
        vendorPlan:
          type: ["string", "null"]
        priority:
          type: string
          enum: [priority, standard, normal]
        status:
          type: string
          enum: [open, closed]
        adminNotes:
          type: ["string", "null"]
        closedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required: [id, name, email, subject, message, priority, status, createdAt]
    CloseSupportTicketBody:
      type: object
      properties:
        adminNotes:
          type: string
    CloseSupportTicketResponse:
      type: object
      properties:
        id:
          type: integer
        status:
          type: string
      required: [id, status]

    # ------------------------------------------------------------------
    # Vendor analytics (plan-gated)
    # ------------------------------------------------------------------
...
    AdminLoginBody:
      type: object
      properties:
        email:
          type: string
          format: email
        password:
          type: string
      required: [email, password]
    AdminUser:
      type: object
      properties:
        id:
          type: integer
        email:
          type: string
        name:
          type: string
      required: [id, email, name]
    AdminOverview:
      type: object
      properties:
        openDisputes:
          type: integer
        awaitingConfirmation:
          type: integer
        openTickets:
          type: integer
        priorityTickets:
          type: integer
        fundsHeldInEscrow:
          type: number
          format: float
      required:
        [openDisputes, awaitingConfirmation, openTickets, priorityTickets, fundsHeldInEscrow]
    DisputeOrderItem:
      type: object
      properties:
        title:
          type: string
        price:
          type: number
          format: float
        itemType:
          type: string
        vendorName:
          type: ["string", "null"]
      required: [title, price]
    DisputeOrder:
      type: object
      properties:
        id:
          type: integer
        total:
          type: number
          format: float
        currency:
          type: string
        status:
          type: string
        paymentStatus:
          type: string
        paymentProvider:
          type: ["string", "null"]
        paymentReference:
          type: ["string", "null"]
        deliveryStatus:
          type: string
        disputeReason:
          type: ["string", "null"]
        disputeRaisedAt:
          type: ["string", "null"]
          format: date-time
        payoutStatus:
          type: string
        refundedAmount:
          type: number
          format: float
        refundableAmount:
          type: number
          format: float
        resolution:
          type: ["string", "null"]
        resolutionNotes:
          type: ["string", "null"]
        resolvedAt:
          type: ["string", "null"]
          format: date-time
        buyerEmail:
          type: ["string", "null"]
        buyerName:
          type: ["string", "null"]
        vendorNames:
          type: array
          items:
            type: string
        items:
          type: array
          items:
            $ref: "#/components/schemas/DisputeOrderItem"
        createdAt:
          type: string
          format: date-time
      required:
        [
          id,
          total,
          currency,
          status,
          paymentStatus,
          deliveryStatus,
          payoutStatus,
          refundedAmount,
          refundableAmount,
          vendorNames,
          items,
          createdAt,
        ]
    ResolveDisputeBody:
      type: object
      properties:
        resolution:
          type: string
          enum: [released, refunded_full, refunded_partial]
        amount:
          type: number
          description: Required for refunded_partial; must be less than the refundable total.
        notes:
          type: string
      required: [resolution]
    ResolveDisputeResponse:
      type: object
      properties:
        orderId:
          type: integer
        resolution:
          type: string
        refundedAmount:
          type: number
          format: float
        refund:
          type: ["object", "null"]
          additionalProperties: true
        payout:
          type: ["object", "null"]
          additionalProperties: true
        resolvedBy:
          type: string
      required: [orderId, resolution, refundedAmount]